feat: Complete Phase 2 - Teacher Space (Course & Module Management)

+++ src/Controller/CourseController.php
@@ -10,8 +10,10 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 #[Route('/course')]
+#[IsGranted('ROLE_TEACHER')]
 final class CourseController extends AbstractController
 {
     #[Route(name: 'app_course_index', methods: ['GET'])]
@@ -54,6 +56,11 @@ final class CourseController extends AbstractController
     #[Route('/{id}/edit', name: 'app_course_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Course $course, EntityManagerInterface $entityManager): Response
     {
+        // Vérifier que c'est bien le cours de cet enseignant
+        if ($course->getTeacher() !== $this->getUser()) {
+            throw $this->createAccessDeniedException("Vous n'êtes pas le propriétaire de ce cours.");
+        }
+
         $form = $this->createForm(CourseType::class, $course);
         $form->handleRequest($request);
 
@@ -72,6 +79,11 @@ final class CourseController extends AbstractController
     #[Route('/{id}', name: 'app_course_delete', methods: ['POST'])]
     public function delete(Request $request, Course $course, EntityManagerInterface $entityManager): Response
     {
+        // Vérifier que c'est bien le cours de cet enseignant
+        if ($course->getTeacher() !== $this->getUser()) {
+            throw $this->createAccessDeniedException("Vous n'êtes pas le propriétaire de ce cours.");
+        }
+
         if ($this->isCsrfTokenValid('delete'.$course->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($course);
             $entityManager->flush();
